+++ Arrow/src/main/java/com/arsyux/arrow/controller/PostController.java
@@ -36,6 +36,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.arsyux.arrow.controller.files.FileService;
 import com.arsyux.arrow.domain.ExhibitionVO;
 import com.arsyux.arrow.domain.FilesVO;
+import com.arsyux.arrow.domain.Pagination;
 import com.arsyux.arrow.dto.ExhibitionDTO;
 import com.arsyux.arrow.dto.ExhibitionDTO.InsertExhibitionValidationGroup;
 import com.arsyux.arrow.dto.ResponseDTO;
@@ -152,13 +153,10 @@ public class PostController {
 
         // 총 페이지 수 계산
         int totalPages = exhibitionService.getTotalPages();
-        System.out.println("totalPages"+totalPages);
-
 		pagination.pageInfo(page, range, totalPages);
         
 		// 페이지 번호를 5개까지만 넘어가게 설정
 		List<ExhibitionVO> contentsList = exhibitionService.selectAllContent(pagination);
-		System.out.println(contentsList);
 
 		model.addAttribute("pagination", pagination);
         model.addAttribute("contentsList", contentsList);
@@ -250,124 +248,7 @@ public class PostController {
 	
 	
 	
-	//20240327 페이징 기능 수정중
-	//이승현
-	public class Pagination {
-		private int listSize = 4;                //초기값으로 목록개수를 10으로 셋팅
-		private int rangeSize = 5;            //초기값으로 페이지범위를 10으로 셋팅
-		private int page;
-		private int range;
-		private int listCnt;
-		private int pageCnt;
-		private int startPage;
-		private int startList;
-		private int endPage;
-		private boolean prev;
-		private boolean next;
-
-		public int getRangeSize() {
-			return rangeSize;
-		}
-
-		public int getPage() {
-			return page;
-		}
-
-		public void setPage(int page) {
-			this.page = page;
-		}
-
-		public int getRange() {
-			return range;
-
-		}
-
-		public void setRange(int range) {
-			this.range = range;
-		}
-
-		public int getStartPage() {
-			return startPage;
-		}
-
-
-		public void setStartPage(int startPage) {
-			this.startPage = startPage;
-		}
-
-		public int getEndPage() {
-			return endPage;
-		}
-
-		public void setEndPage(int endPage) {
-			this.endPage = endPage;
-		}
-
-		public boolean isPrev() {
-			return prev;
-		}
-
-		public void setPrev(boolean prev) {
-			this.prev = prev;
-		}
-
-		public boolean isNext() {
-			return next;
-		}
-
-		public void setNext(boolean next) {
-			this.next = next;
-		}
-
-		public int getListSize() {
-			return listSize;
-		}
-
-		public void setListSize(int listSize) {
-			this.listSize = listSize;
-		}
-
-		public int getListCnt() {
-			return listCnt;
-		}
-
-		public void setListCnt(int listCnt) {
-			this.listCnt = listCnt;
-		}
-
-		public int getStartList() {
-			return startList;
-		}
-
-		public void pageInfo(int page, int range, int listCnt) {
-			this.page = page;
-			this.range = range;
-			this.listCnt = listCnt;
-
-			//전체 페이지수 
-			this.pageCnt = (int) Math.ceil(listCnt/listSize);			
-
-			//시작 페이지
-			this.startPage = (range - 1) * rangeSize + 1 ;
-
-			//끝 페이지
-			this.endPage = range * rangeSize;
-
-			//게시판 시작번호
-			this.startList = (page - 1) * listSize;
-
-			//이전 버튼 상태
-			this.prev = range == 1 ? false : true;
-
-			//다음 버튼 상태
-			this.next = endPage > pageCnt ? false : true;
-			
-			if (this.endPage > this.pageCnt) {
-				this.endPage = this.pageCnt;
-				this.next = false;
-			}
-		}
-	}
+	
 	
 	
 }
