+++ Arrow/src/main/java/com/arsyux/arrow/service/CollectionService.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.arsyux.arrow.domain.CollectionsVO;
 import com.arsyux.arrow.domain.ExhibitionVO;
 import com.arsyux.arrow.domain.FilesVO;
+import com.arsyux.arrow.domain.Pagination;
 import com.arsyux.arrow.persistence.CollectionDAO;
 import com.arsyux.arrow.persistence.ExhibitionDAO;
 
@@ -27,16 +28,16 @@ public class CollectionService {
 	
 
 
-	public List<CollectionsVO> selectCollect(int exh_seq,int pageNumber, int pageSize) {
-		int offset = pageNumber * pageSize;
-		return collectionDAO.selectCollect(exh_seq,offset, pageSize);
+	public List<CollectionsVO> selectCollect(int exh_seq,Pagination page) {
+		int startList = page.getStartList();
+		int listSize = page.getListSize();
+		RowBounds rowBounds = new RowBounds(startList, listSize);
+		return collectionDAO.selectCollect(exh_seq, rowBounds);
 	}
 
-    public int getTotalPages(int pageSize) {
-        // 총 게시글 수를 가져와서 페이지 수 계산
-        int totalContents = collectionDAO.getTotalPages(pageSize);
-        int totalPages = (int) Math.ceil((double) totalContents / pageSize);
-        return totalPages;
+    public int getTotalPages() {      
+    
+        return collectionDAO.getTotalPages();
     }
